fixes to remove warnings from gcc-4.0.0

+++ src/plugin_common/charset.c
@@ -88,7 +88,7 @@ char* FLAC_plugin__charset_convert_string (const char *string, char *from, char 
 	outptr = out;
 
 retry:
-	if (iconv(cd, (char**)&input, &length, &outptr, &outleft) == -1)
+	if (iconv(cd, (char**)&input, &length, &outptr, &outleft) == (size_t)(-1))
 	{
 		int used;
 		switch (errno)
